feat: split create_bulk_items into create and update tools
Separates the single bulk tool into two distinct tools so LLMs can
better recognize update capabilities:
- create_bulk_items: creates new items (group_id optional, defaults to "topics")
- update_bulk_items: updates existing items via on_match (group_id required)

Shared logic extracted to bulk-items.utils.ts.
