+++ packages/agent-toolkit/src/core/tools/platform-api-tools/bulk-items-tools/bulk-items.graphql.ts
