+++ packages/agent-toolkit/src/core/tools/platform-api-tools/bulk-items-tools/index.ts
@@ -1,0 +1,2 @@
+export { CreateBulkItemsTool } from './create-bulk-items-tool';
+export { UpdateBulkItemsTool } from './update-bulk-items-tool';
