+++ packages/agent-toolkit/src/core/tools/platform-api-tools/index.ts
@@ -72,7 +72,7 @@ import { CreateAgentTool } from './agents-tools/create-agent/create-agent-tool';
 import { DeleteAgentTool } from './agents-tools/delete-agent/delete-agent-tool';
 import { ListWorkflowsTool } from './workflows-tools/list-workflows/list-workflows-tool';
 import { ManageWorkflowsTool } from './workflows-tools/manage-workflows/manage-workflows-tool';
-import { CreateBulkItemsTool } from './create-bulk-items-tool/create-bulk-items-tool';
+import { CreateBulkItemsTool, UpdateBulkItemsTool } from './bulk-items-tools';
 
 export const allGraphqlApiTools: BaseMondayApiToolConstructor[] = [
   DeleteItemTool,
@@ -153,6 +153,7 @@ export const allGraphqlApiTools: BaseMondayApiToolConstructor[] = [
   ManageWorkflowsTool,
   // Bulk operations
   CreateBulkItemsTool,
+  UpdateBulkItemsTool,
 ];
 
 export * from './all-monday-api-tool';
@@ -228,6 +229,6 @@ export * from './workflows-tools';
 // Dashboard Tools
 export * from './dashboard-tools';
 // Bulk operations
-export * from './create-bulk-items-tool/create-bulk-items-tool';
+export * from './bulk-items-tools';
 // Monday Dev Tools
 export * from '../monday-dev-tools';
